requirements.txt
torch>=2.0
tokenizers>=0.13.3
transformers
datasets
streamlit
fpdf
python-docx
pyyaml
